Fix incongruence between readme and compose.yaml
Previously the compose.yaml file was reading from a cron/ddns-cloudflare subdirectory, whereas the instructions in README.md referred to the host's /etc/opt/ddns-cloudflare.

This was not noticed because in the dev environment the former was symlinked to the latter. This symlink is now included in the repository and future releases.